Move ASM generation to a different class. This prevents GraalVM to load ASM classes even when substituting getBytes.
